Added tests and ros bag updates
