+++ TBD.Psi.RosBagStreamReader/RosStreamMetaData.cs
@@ -5,7 +5,7 @@ namespace TBD.Psi.RosBagStreamReader
     public class RosStreamMetaData : StreamMetadataBase
     {
         public RosStreamMetaData(string name, int sourceId, string typeName, string partitionName, string partitionPath, DateTime first, DateTime last, int averageMessageSize, int messageCount, int averageLatency)
-                : base(name, sourceId, typeName, partitionName, partitionPath, first, last, averageMessageSize, messageCount, averageLatency)
+                : base(name, sourceId, typeName, partitionName, partitionPath, first, last, messageCount, averageMessageSize, averageLatency)
             {
             this.CreationTimeInterval = new TimeInterval(first, last);
             }
