+++ TBD.Psi.TransformationTree/TBD.Psi.TransformationTree.csproj
@@ -2,7 +2,23 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <Version>0.1.0-beta</Version>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <Company>Carnegie Mellon University</Company>
+    <Authors>Transportation, Bots, and Disability Lab</Authors>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <RepositoryUrl>https://github.com/CMU-TBD/TBD.Psi.Components</RepositoryUrl>
+    <Copyright>Transportation, Bots, and Disability Lab - Carnegie Mellon University</Copyright>
+    <Description>A Psi-hybrid component that enable spatial transformations queries through transformation chaining.</Description>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
+
+  <ItemGroup>
+    <None Include="..\LICENSE">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="MathNet.Spatial.Signed" Version="0.6.0" />
